% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/print_hello_rmedicine.R
\name{print_hello_rmedicine}
\alias{print_hello_rmedicine}
\title{Print hello R/Medicine}
\usage{
print_hello_rmedicine(x)
}
\arguments{
\item{x}{This argument does nothing.
If it did do something, it might require more words to explain.}
}
\value{
Prints \code{"Hello, R/Medicine"} to the console
}
\description{
This function just prints hello R/Medicine.
}
\examples{
print_hello_rmedicine(x = 1)

}
