% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/print_hello_world.R
\name{print_hello_world}
\alias{print_hello_world}
\title{Print hello world}
\usage{
print_hello_world(x)
}
\arguments{
\item{x}{This argument does nothing.
If it did do something, it might require more words to explain.}
}
\value{
Prints \code{"Hello, world!"} to the console
}
\description{
This function just prints hello world.
}
\examples{
print_hello_world(x = 1)

}
